refactor: changed all model names to have model as suffix
Also created `AuthenticationCodeModel` migrations and refactored `UserTokenModel` and `UserTokenModelMigrations` to their own separate folders - no longer coupled with `UserModel`
