+++ .gitignore
@@ -14,7 +14,7 @@ db.sqlite
 #
 # gitignore contributors: remember to update Global/Xcode.gitignore, Objective-C.gitignore & Swift.gitignore
 
-## User settings
+## UserModel settings
 xcuserdata/
 
 ## Obj-C/Swift specific
