+++ Sources/App/Controllers/BadgeController/BadgeController.swift
@@ -15,12 +15,15 @@ struct BadgeController: RouteCollection {
 
     @Sendable
     func getAllBadges(req: Request) async throws -> [BadgeDTO.GetBadge] {
-        try await Badge.query(on: req.db).all().map { $0.toDTO() }
+        try await BadgeModel.query(on: req.db).all().map { $0.toDTO() }
     }
 
     @Sendable
     func getBadgeInfo(req: Request) async throws -> BadgeDTO.GetBadge {
-        guard let badge = try await Badge.find(req.parameters.get("badgeID", as: UUID.self), on: req.db) else {
+        guard
+            let badge = try await BadgeModel.find(
+                req.parameters.get("badgeID", as: UUID.self), on: req.db)
+        else {
             throw Abort(.notFound)
         }
         return badge.toDTO()
