+++ Sources/App/Services/AuthenticationService/AuthenticationService.swift
@@ -6,7 +6,7 @@ struct AuthenticationService {
 
     // TODO: add some two step verification to verify the email belongs to the creator
 
-    func sendLoginCode(email: String) async throws
+    func sendLoginCode(email: String) async throws {
         // 1. Do some rate limit checks, e.g. user may only send a code once every 5 minutes and a total of 10 in a 24 hour time period
         // - If user reaches rate limit, send appropriate response like "You sent too many two factor authentication codes today. Try again tomorrow" or "try again in an hour"
         // - Send an email to the user once they can send a code again to login.
@@ -17,9 +17,9 @@ struct AuthenticationService {
         // 2. Send login code to user email
         // - save code to database (so that the rate limit checks above will work accurately)
         // - code must have an expiration date, meaning it will only be valid for let's say 5 minutes
+    }
 
-
-    func login(user: User) async throws -> String {
+    func login(user: UserModel) async throws -> String {
         let token = try self.generateBearerToken(for: user)
         try await token.save(on: db)
         return token.value
@@ -28,18 +28,18 @@ struct AuthenticationService {
     func logout(user userID: UUID, req: Request) async throws {
         // TODO: send messages to message service here
         try await self.removeOldBearerTokens(for: userID)
-        req.auth.logout(User.self)
+        req.auth.logout(UserModel.self)
 
     }
 
-    func generateBearerToken(for user: User) throws -> UserToken {
+    func generateBearerToken(for user: UserModel) throws -> UserTokenModel {
         try .init(
             value: [UInt8].random(count: 16).base64,
             userID: user.requireID()
         )
     }
 
-    func getUserFromBearerAuthorization(_ bearer: BearerAuthorization) async throws -> User {
+    func getUserFromBearerAuthorization(_ bearer: BearerAuthorization) async throws -> UserModel {
         // 1. Get token from db
         let token = try await getBearerToken(bearer.token)
 
@@ -52,11 +52,11 @@ struct AuthenticationService {
         return try await token.$user.get(on: db)
     }
 
-    func getUserFromBasicAuthorization(_ basic: BasicAuthorization) async throws -> User {
+    func getUserFromBasicAuthorization(_ basic: BasicAuthorization) async throws -> UserModel {
         // 1. Get user from db
         guard
             let user =
-                try await User
+                try await UserModel
                 .query(on: db)
                 .filter(\.$username, .equal, basic.username)
                 .first()
@@ -75,10 +75,10 @@ struct AuthenticationService {
         }
     }
 
-    func getBearerToken(_ token: String) async throws -> UserToken {
+    func getBearerToken(_ token: String) async throws -> UserTokenModel {
         guard
             let userToken =
-                try await UserToken
+                try await UserTokenModel
                 .query(on: db)
                 .filter(\.$value, .equal, token)
                 .first()
@@ -90,7 +90,7 @@ struct AuthenticationService {
 
     func removeOldBearerTokens(for userID: UUID) async throws {
         let tokens =
-            try await UserToken
+            try await UserTokenModel
             .query(on: db)
             .filter(\.$user.$id, .equal, userID)
             .all()
